+++ Documentation/Changelog.md
@@ -13,6 +13,7 @@ Release date: not released yet
 - Added `Fn.Ignore(value)` function that ignores a specified value for languages that do not have a built-in equivalent. A typical usage is to fire and forget a parallel `Task` without producing a compiler warning
 - Added `Process.GetImageFileName()` method that allows to retrieve file name of a running process without security restrictions imposed by the host OS
 - Added `bool ISet.AddRange(collection)` extension method that allows to add elements to a set in bulk
+- Added `CommandLine.EscapeArgument` and `CommandLine.EscapeFileName` methods
 - .NET Framework 4.0 target is retired. The minimal supported .NET Framework version is 4.5
 - Fixed issue with ambiguous match of `IsNullOrEmpty` polyfill method of `HashSet<T>` type that occurred in .NET 4.6+, .NET Standard 2.0+ and .NET Core 2.0+ target frameworks
 - Fixed issue with ambiguous match of `ToHashSet` polyfill method of `IEnumerable<T>` type that occurred in .NET 4.7.2+ target frameworks